% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_BuildBackend.R
\name{BuildBackend}
\alias{BuildBackend}
\title{Build the environment for CausalImages models. Builds a conda environment in which jax, tensorflow, tensorflow-probability optax, equinox, and jmp are installed.}
\usage{
BuildBackend(conda_env = "CausalImagesEnv", conda = "auto")
}
\arguments{
\item{conda_env}{(default = \code{"CausalImagesEnv"}) Name of the conda environment in which to place the backends.}

\item{conda}{(default = \code{auto}) The path to a conda executable. Using \code{"auto"} allows reticulate to attempt to automatically find an appropriate conda binary.}
}
\value{
Builds the computational environment for \code{causalimages}. This function requires an Internet connection.
You may find out a list of conda Python paths via: \code{system("which python")}
}
\description{
Build the environment for CausalImages models. Builds a conda environment in which jax, tensorflow, tensorflow-probability optax, equinox, and jmp are installed.
}
\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
