% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_GetAndSaveGeolocatedImages.R
\name{GetAndSaveGeolocatedImages}
\alias{GetAndSaveGeolocatedImages}
\title{Getting and saving geo-located images from a pool of .tif's}
\usage{
GetAndSaveGeolocatedImages(
  long,
  lat,
  keys,
  tif_pool,
  image_pixel_width = 250L,
  save_folder = ".",
  save_as = "csv",
  lyrs = NULL
)
}
\arguments{
\item{long}{Vector of numeric longitudes.}

\item{lat}{Vector of numeric latitudes.}

\item{keys}{The image keys associated with the long/lat coordinates.}

\item{tif_pool}{A character vector specifying the fully qualified path to a corpus of .tif files.}

\item{image_pixel_width}{An even integer specifying the pixel width (and height) of the saved images.}

\item{save_folder}{(default = \code{"."}) What folder should be used to save the output? Example: \code{"~/Downloads"}}

\item{save_as}{(default = \code{".csv"}) What format should the output be saved as? Only one option currently (\code{.csv})}

\item{lyrs}{(default = NULL) Integer (vector) specifying the layers to be extracted. Default is for all layers to be extracted.}
}
\value{
Finds the image slice associated with the \code{long} and \code{lat} values, saves images by band (if \code{save_as = "csv"}) in save_folder.
The save format is: \code{sprintf("\%s/Key\%s_BAND\%s.csv", save_folder, keys[i], band_)}
}
\description{
A function that finds the image slice associated with the \code{long} and \code{lat} values, saves images by band (if \code{save_as = "csv"}) in save_folder.
}
\examples{

# Example use (not run):
#MASTER_IMAGE_POOL_FULL_DIR <- c("./LargeTifs/tif1.tif","./LargeTifs/tif2.tif")
#GetAndSaveGeolocatedImages(
                       #long = GeoKeyMat$geo_long,
                       #lat = GeoKeyMat$geo_lat,
                       #image_pixel_width = 500L,
                       #keys = row.names(GeoKeyMat),
                       #tif_pool = MASTER_IMAGE_POOL_FULL_DIR,
                       #save_folder = "./Data/Uganda2000_processed",
                       #save_as = "csv",
                       #lyrs = NULL)

}
