% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_TfRecordFxns.R
\name{GetElementFromTfRecordAtIndices}
\alias{GetElementFromTfRecordAtIndices}
\title{Reads unique key indices from a \code{.tfrecord} file.}
\usage{
GetElementFromTfRecordAtIndices(
  uniqueKeyIndices,
  filename,
  nObs,
  readVideo = F,
  conda_env = NULL,
  conda_env_required = F,
  image_dtype = "float16",
  iterator = NULL,
  return_iterator = F
)
}
\arguments{
\item{uniqueKeyIndices}{(integer vector) Unique image indices to be retrieved from a \code{.tfrecord}}

\item{filename}{(character string) A character string stating the path to a \code{.tfrecord}}

\item{nObs}{(integer) Total number of observations in the tfrecord file}

\item{readVideo}{(default = \code{FALSE}) A Boolean indicating whether to read video/sequence data}

\item{conda_env}{(Default = \code{NULL}) A \code{conda} environment where tensorflow v2 lives. Used only if a version of tensorflow is not already active.}

\item{conda_env_required}{(default = \code{FALSE}) A Boolean stating whether use of the specified conda environment is required.}

\item{image_dtype}{(default = \code{"float16"}) A string specifying the image data type}

\item{iterator}{(default = \code{NULL}) An optional iterator object for sequential reading}

\item{return_iterator}{(default = \code{FALSE}) A Boolean indicating whether to return the iterator for subsequent calls}
}
\value{
Returns content from a \code{.tfrecord} associated with \code{uniqueKeyIndices}
}
\description{
Reads unique key indices from a \code{.tfrecord} file saved via a call to \code{causalimages::WriteTfRecord}.
}
\examples{
# Example usage (not run):
#GetElementFromTfRecordAtIndices(
  #uniqueKeyIndices = 1:10,
  #filename = "./NigeriaConfoundApp.tfrecord",
  #nObs = 100)

}
