% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_GetMoments.R
\name{GetMoments}
\alias{GetMoments}
\title{Get moments for normalization (internal function)}
\usage{
GetMoments(iterator, dataType, image_dtype, momentCalIters = 34L)
}
\arguments{
\item{iterator}{An iterator}

\item{dataType}{A string denoting data type}

\item{image_dtype}{A string specifying the image data type (e.g., "float16", "float32")}

\item{momentCalIters}{Number of minibatches with which to estimate moments}
}
\value{
Returns mean/sd arrays for normalization.
}
\description{
An internal function for obtaining moments for channel normalization.
}
\examples{
# (Not run)
# GetMoments(iterator, dataType, image_dtype, momentCalIters = 34L)
}
