% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_RandomizedEmbedding.R
\name{GetRandomizedImageEmbeddings}
\alias{GetRandomizedImageEmbeddings}
\title{Generates randomized image and video embeddings useful in earth observation tasks for casual inference, generalizing the approach in Rolf, Esther, et al.  (2021).}
\usage{
GetRandomizedImageEmbeddings(imageKeysOfUnits, acquireImageFxn, nFeatures, ...)
}
\arguments{
\item{acquireImageFxn}{A function specifying how to load images representations associated with \code{imageKeysOfUnits} into memory. For example, if observation \code{3} has a value  of \code{"a34f"} in \code{imageKeysOfUnits}, \code{acquireImageFxn} should extract the image associated with the unique key \code{"a34f"}.
First argument should be image key values and second argument have be \code{training} (in case different behavior in training/inference mode).}

\item{imageKeysOfUnits}{A vector of length \code{length(imageKeysOfUnits)} specifying the unique image ID associated with each unit. Samples of \code{imageKeysOfUnits} are fed into \code{acquireImageFxn} to call images into memory.}

\item{conda_env}{(default = \code{NULL}) A string specifying a conda environment wherein \code{tensorflow}, \code{tensorflow_probability}, and \code{gc} are installed.}

\item{conda_env_required}{(default = \code{F}) A Boolean stating whether use of the specified conda environment is required.}

\item{nFeatures}{(default = \code{256L}) Dimensions used in the convolution kernels.}

\item{batchSize}{(default = \code{50L}) Integer specifying batch size in obtaining embeddings.}

\item{strides}{(default = \code{2L}) Integer specifying the strides used in the convolutional layers.}

\item{temporalKernelSize}{(default = \code{2L}) Dimensions used in the temporal part of the convolution kernels if using image sequences.}

\item{kernelSize}{(default = \code{5L}) Dimensions used in the convolution kernels.}

\item{seed}{(default = \code{NULL}) Integer specifying the seed.}

\item{outputType}{(default = \code{"R"}) Either \code{"R"} or \code{"tensorflow"} indicating whether to output R or tensorflow arrays.}
}
\value{
A list containing two items:
\itemize{
\item \code{embeddings} (matrix) A matrix containingimage/video embeddings, with rows corresponding to observations.
\item \code{embeddings_fxn} (function) The functioning performing the embedding, returned for re-use.
}
}
\description{
Generates randomized image and video embeddings useful in earth observation tasks for casual inference, generalizing the approach in Rolf, Esther, et al.  (2021).
}
\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/
 @section References:
\itemize{
\item Rolf, Esther, et al. "A generalizable and accessible approach to machine learning with global satellite imagery." *Nature Communications* 12.1 (2021): 4392.
}

}
