% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_TFRecordManagement.R
\name{TFRecordManagement}
\alias{TFRecordManagement}
\title{Defines an internal TFRecord management routine (internal function)}
\usage{
TFRecordManagement()
}
\arguments{
\item{.}{No parameters.}
}
\value{
Internal function defining a tfrecord management sequence.
}
\description{
Defines management defined in TFRecordManagement(). Internal function.
}
