% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_TrainDefine.R
\name{TrainDefine}
\alias{TrainDefine}
\title{Defines an internal training routine (internal function)}
\usage{
TrainDefine()
}
\arguments{
\item{.}{No parameters.}
}
\value{
Internal function defining a training sequence.
}
\description{
Defines trainers defined in TrainDefine(). Internal function.
}
