% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Confounding.R
\name{AnalyzeImageConfounding}
\alias{AnalyzeImageConfounding}
\title{Perform causal estimation under image confounding}
\usage{
AnalyzeImageConfounding(
  obsW,
  obsY,
  X = NULL,
  file = NULL,
  imageKeysOfUnits = NULL,
  fileTransport = NULL,
  imageKeysOfUnitsTransport = NULL,
  nBoot = 10L,
  inputAvePoolingSize = 1L,
  useTrainingPertubations = T,
  useScalePertubations = F,
  orthogonalize = F,
  transportabilityMat = NULL,
  latTransport = NULL,
  longTransport = NULL,
  lat = NULL,
  long = NULL,
  conda_env = "CausalImagesEnv",
  conda_env_required = T,
  Sys.setenv_text = NULL,
  figuresTag = NULL,
  figuresPath = "./",
  plotBands = 1L,
  plotResults = T,
  XCrossModal = T,
  XForceModal = F,
  optimizeImageRep = T,
  nonLinearScaler = NULL,
  nWidth_ImageRep = 64L,
  nDepth_ImageRep = 1L,
  kernelSize = 5L,
  nWidth_Dense = 64L,
  nDepth_Dense = 1L,
  imageModelClass = "VisionTransformer",
  pretrainedModel = NULL,
  strides = 2L,
  nDepth_TemporalRep = 3L,
  patchEmbedDim = 16L,
  dropoutRate = 0.1,
  droppathRate = 0.1,
  batchSize = 16L,
  nSGD = 400L,
  testFrac = 0.05,
  TfRecords_BufferScaler = 4L,
  learningRateMax = 0.001,
  TFRecordControl = NULL,
  dataType = "image",
  image_dtype = "float16",
  atError = "stop",
  seed = NULL
)
}
\arguments{
\item{obsW}{A numeric vector where \code{0}'s correspond to control units and \code{1}'s to treated units.}

\item{obsY}{A numeric vector containing observed outcomes.}

\item{X}{An optional numeric matrix containing tabular information used if \code{orthogonalize = T}. \code{X} is normalized internally and salience maps with respect to \code{X} are transformed back to the original scale.}

\item{file}{Path to a tfrecord file generated by \code{WriteTfRecord}.}

\item{imageKeysOfUnits}{A vector of length \code{length(obsY)} specifying the unique image ID associated with each unit. Samples of \code{imageKeysOfUnits} are fed into the package to call images into memory.}

\item{nBoot}{Number of bootstrap iterations for uncertainty estimation.}

\item{useTrainingPertubations}{Boolean specifying whether to randomly perturb the image axes during training to reduce overfitting.}

\item{transportabilityMat}{Optional matrix with a column named \code{imageKeysOfUnits} specifying keys to be used by the package for generating treatment effect predictions for out-of-sample points.}

\item{long, lat}{Optional vectors specifying longitude and latitude coordinates for units. Used only for describing highest and lowest probability neighborhood units if specified.}

\item{conda_env}{A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}. Default = \code{"CausalImagesEnv"}.}

\item{conda_env_required}{A Boolean stating whether use of the specified conda environment is required.}

\item{figuresTag}{A string specifying an identifier that is appended to all figure names.}

\item{figuresPath}{A string specifying file path for saved figures made in the analysis.}

\item{plotBands}{An integer or vector specifying which band position (from the image representation) should be plotted in the visual results. If a vector, \code{plotBands} should have 3 (and only 3) dimensions (corresponding to the 3 dimensions to be used in RGB plotting).}

\item{plotResults}{(default = \code{T}) Should analysis results be plotted?}

\item{optimizeImageRep}{Boolean specifying whether to optimize over the image model representation (or only over downstream parameters).}

\item{nWidth_ImageRep}{Integer specifying width of image model representation.}

\item{nDepth_ImageRep}{Integer specifying depth of image model representation.}

\item{kernelSize}{Dimensions used in spatial convolutions.}

\item{nWidth_Dense}{Integer specifying width of image model representation.}

\item{nDepth_Dense}{Integer specifying depth of dense model representation.}

\item{strides}{(default = \code{2L}) Integer specifying the strides used in the convolutional layers.}

\item{dropoutRate}{Dropout rate used in training to prevent overfitting (\code{dropoutRate = 0} corresponds to no dropout).}

\item{droppathRate}{Droppath rate used in training to prevent overfitting (\code{droppathRate = 0} corresponds to no droppath).}

\item{batchSize}{Batch size used in SGD optimization. Default = \code{50L}.}

\item{nSGD}{Number of stochastic gradient descent (SGD) iterations. Default = \code{400L}}

\item{testFrac}{Default = \code{0.1}. Fraction of observations held out as a test set to evaluate out-of-sample loss values.}

\item{TfRecords_BufferScaler}{The buffer size used in \code{tfrecords} mode is \code{batchSize*TfRecords_BufferScaler}. Lower \code{TfRecords_BufferScaler} towards 1 if out-of-memory problems.}

\item{dataType}{(default = \code{"image"}) String specifying whether to assume \code{"image"} or \code{"video"} data types.}
}
\value{
Returns a list consisting of
\itemize{
\item \code{ATE_est} ATE estimate.
\item \code{ATE_se} Standard error estimate for the ATE.
\item \code{plotResults} If set to \code{TRUE}, causal salience plots are saved to disk, characterizing the image confounding structure. See references for details.
}
}
\description{
Perform causal estimation under image confounding
}
\section{References}{

\itemize{
\item  Connor T. Jerzak, Fredrik Johansson, Adel Daoud. Integrating Earth Observation Data into Causal Inference: Challenges and Opportunities. \emph{ArXiv Preprint}, 2023.
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
