% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_Confounding.R
\name{AnalyzeImageConfounding}
\alias{AnalyzeImageConfounding}
\title{Perform causal estimation under image confounding}
\usage{
AnalyzeImageConfounding(obsW, obsY, imageKeysOfUnits, file, ...)
}
\arguments{
\item{obsW}{A numeric vector where \code{0}'s correspond to control units and \code{1}'s to treated units.}

\item{obsY}{A numeric vector containing observed outcomes.}

\item{X}{(default = \code{NULL}) An optional numeric matrix containing tabular information used if \code{orthogonalize = T}. \code{X} is normalized internally and salience maps with respect to \code{X} are transformed back to the original scale.}

\item{file}{(default = \code{NULL}) Path to a tfrecord file generated by \code{WriteTfRecord}.}

\item{imageKeysOfUnits}{(default = \code{1:length(obsY)}) A vector of length \code{length(obsY)} specifying the unique image ID associated with each unit. Samples of \code{imageKeysOfUnits} are fed into the package to call images into memory.}

\item{nBoot}{(default = \code{100L}) Number of bootstrap iterations for uncertainty estimation.}

\item{transportabilityMat}{(optional) A matrix with a column named \code{imageKeysOfUnits} specifying keys to be used by the package for generating treatment effect predictions for out-of-sample points.}

\item{long, lat}{(optional) Vectors specifying longitude and latitude coordinates for units. Used only for describing highest and lowest probability neighorhood units if specified.}

\item{conda_env}{(default = \code{"CausalImagesEnv"}) A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}}

\item{conda_env_required}{(default = \code{T}) A Boolean stating whether use of the specified conda environment is required.}

\item{figuresTag}{(default = \code{""}) A string specifying an identifier that is appended to all figure names.}

\item{figuresPath}{(default = \code{"./"}) A string specifying file path for saved figures made in the analysis.}

\item{plotBands}{(default = \code{1L}) An integer or vector specifying which band position (from the image representation) should be plotted in the visual results. If a vector, \code{plotBands} should have 3 (and only 3) dimensions (corresponding to the 3 dimensions to be used in RBG plotting).}

\item{plotResults}{(default = \code{T}) Should analysis results be plotted?}

\item{strides}{(default = \code{2L}) Integer specifying the strides used in the convolutional layers.}

\item{dropoutRate}{(default = \code{0.1}) Droppout rate used in training used to prevent overfitting (\code{dropoutRate = 0} corresponds to no dropout).}

\item{batchSize}{(default = \code{50L}) Batch size used in SGD optimization.}

\item{kernelSize}{(default = \code{5L}) Dimensions used in convolution kernels.}

\item{nSGD}{(default = \code{400L}) Number of stochastic gradient descent (SGD) iterations.}

\item{testFrac}{(default = \code{0.1}) Fraction of observations held out as a test set to evaluate out-of-sample loss values.}

\item{channelNormalize}{(default = \code{T}) Should channelwise image feature normalization be attempted? Default is \code{T}, as this improves training.}

\item{TfRecords_BufferScaler}{(default = \code{4L}) The buffer size used in \code{tfrecords} mode is \code{batchSize*TfRecords_BufferScaler}. Lower \code{TfRecords_BufferScaler} towards 1 if out-of-memory problems.}
}
\value{
Returns a list consisting of
\itemize{
\item \code{ATE_est} ATE estimate.
\item \code{ATE_se} Standard error estimate for the ATE.
\item \code{plotResults} If set to \code{TRUE}, causal salience plots are saved to disk, characterizing the image confounding structure. See references for details.
}
}
\description{
Perform causal estimation under image confounding
}
\section{References}{

\itemize{
\item  Connor T. Jerzak, Fredrik Johansson, Adel Daoud. Integrating Earth Observation Data into Causal Inference: Challenges and Opportunities. \emph{ArXiv Preprint}, 2023.
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
