% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_Heterogeneity.R
\name{AnalyzeImageHeterogeneity}
\alias{AnalyzeImageHeterogeneity}
\title{AnalyzeImageHeterogeneity}
\usage{
AnalyzeImageHeterogeneity(
  obsW,
  obsY,
  X = NULL,
  imageKeys = NULL,
  transportabilityMat = NULL,
  lat = NULL,
  long = NULL,
  externalFigureKey = "",
  acquireImageRepFxn,
  acquireImageFxn_full = acquireImageRepFxn,
  TYPE = "variational_minimal",
  SimMode = F,
  nDepth_conv = 1,
  nDepth_dense = 1,
  plotResults = F,
  figuresPath = "./",
  kClust_est = 2,
  maxPoolSize = 2L,
  strides = 1L,
  nMonte_predictive = 10L,
  y_density = "normal",
  orthogonalize = F,
  compile = F,
  nMonte_variational = 5L,
  kernelWidth,
  nSGD = 400,
  nDenseWidth = 64L,
  nFilters = 7L
)
}
\arguments{
\item{DAG}{'DAG'.}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements ...
}
\section{References}{

\itemize{
\item References here
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
