% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_Heterogeneity.R
\name{AnalyzeImageHeterogeneity}
\alias{AnalyzeImageHeterogeneity}
\title{Decompose treatment effect heterogeneity by image}
\usage{
AnalyzeImageHeterogeneity(obsW, obsY, acquireImageFxn, ...)
}
\arguments{
\item{obsW}{'DAG'.}

\item{nMonte_predictive}{(default = `10L`) An integer specifying how many Monte Carlo iterations to use in the calculation
approximate posterior means (e.g., mean cluster probabilities).}
}
\value{
A list consiting of \itemize{
  \item Items.
}
}
\description{
Implements the image heterogeneity decomposition analysis of Jerzak, Johansson, and Daoud (2023).
}
\section{References}{

\itemize{
\item Connor T. Jerzak, Fredrik Johansson, Adel Daoud. Image-based Treatment Effect Heterogeneity. Forthcoming in \emph{Proceedings of the Second Conference on Causal Learning and Reasoning (CLeaR), Proceedings of Machine Learning Research (PMLR)}, 2023.
}
}

\examples{
#set seed
set.seed(1)

#Geneate data
x <- rnorm(100)

}
