% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_Heterogeneity.R
\name{AnalyzeImageHeterogeneity}
\alias{AnalyzeImageHeterogeneity}
\title{Decompose treatment effect heterogeneity by image or image sequence}
\usage{
AnalyzeImageHeterogeneity(
  obsW,
  obsY,
  X = NULL,
  orthogonalize = F,
  imageKeysOfUnits = 1:length(obsY),
  kClust_est = 2,
  file = NULL,
  transportabilityMat = NULL,
  lat = NULL,
  long = NULL,
  conda_env = "CausalImagesEnv",
  conda_env_required = T,
  figuresTag = "",
  figuresPath = "./",
  plotBands = 1L,
  heterogeneityModelType = "variational_minimal",
  plotResults = F,
  optimizeImageRep = T,
  nWidth_ImageRep = 64L,
  nDepth_ImageRep = 1L,
  nWidth_Dense = 64L,
  nDepth_Dense = 1L,
  nDepth_TemporalRep = 1L,
  useTrainingPertubations = T,
  strides = 2L,
  pretrainedModel = NULL,
  testFrac = 0.1,
  kernelSize = 5L,
  learningRateMax = 0.005,
  patchEmbedDim = 16L,
  nSGD = 500L,
  batchSize = 16L,
  seed = NULL,
  Sys.setenv_text = NULL,
  imageModelClass = "VisionTransformer",
  nMonte_predictive = 10L,
  nMonte_salience = 10L,
  nMonte_variational = 2L,
  TfRecords_BufferScaler = 4L,
  temperature = 1,
  inputAvePoolingSize = 1L,
  dataType = "image"
)
}
\arguments{
\item{obsW}{A numeric vector where \code{0}'s correspond to control units and \code{1}'s to treated units.}

\item{obsY}{A numeric vector containing observed outcomes.}

\item{X}{Optimal numeric matrix containing tabular information used if \code{orthogonalize = T}.}

\item{orthogonalize}{A Boolean specifying whether to perform the image decomposition after orthogonalizing with respect to tabular covariates specified in \code{X}.}

\item{imageKeysOfUnits}{A vector of length \code{length(obsY)} specifying the unique image ID associated with each unit. Samples of \code{imageKeysOfUnits} are fed into the package to call images into memory.}

\item{kClust_est}{Integer specifying the number of clusters used in estimation. Default is \code{2L}.}

\item{file}{Path to a tfrecord file generated by \code{WriteTfRecord}.}

\item{transportabilityMat}{An optional matrix with a column named \code{key} specifying keys to be used for generating treatment effect predictions for out-of-sample points in earth observation data settings.}

\item{long, lat}{Optional vectors specifying longitude and latitude coordinates for units. Used only for describing highest and lowest probability neighorhood units if specified.}

\item{conda_env}{A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}. Default = \code{"CausalImagesEnv"}.}

\item{conda_env_required}{A Boolean stating whether use of the specified conda environment is required.}

\item{figuresTag}{A string specifying an identifier that is appended to all figure names.}

\item{figuresPath}{A string specifying file path for saved figures made in the analysis.}

\item{plotBands}{An integer or vector specifying which band position (from the acquired image representation) should be plotted in the visual results. If a vector, \code{plotBands} should have 3 (and only 3) dimensions (corresponding to the 3 dimensions to be used in RGB plotting).}

\item{plotResults}{Should analysis results be plotted?}

\item{optimizeImageRep}{Boolean specifying whether to optimize over the image model representation (or only over downstream parameters).}

\item{nWidth_ImageRep}{Integer specifying width of image model representation.}

\item{nDepth_ImageRep}{Integer specifying depth of image model representation.}

\item{nWidth_Dense}{Integer specifying width of image model representation.}

\item{nDepth_Dense}{Integer specifying depth of dense model representation.}

\item{strides}{Integer specifying the strides used in the convolutional layers.=}

\item{kernelSize}{Dimensions used in spatial convolutions.}

\item{nSGD}{Number of stochastic gradient descent (SGD) iterations.}

\item{batchSize}{Batch size used in SGD optimization.}

\item{nMonte_predictive}{An integer specifying how many Monte Carlo iterations to use in the calculation
of posterior means (e.g., mean cluster probabilities).}

\item{nMonte_salience}{An integer specifying how many Monte Carlo iterations to use in the calculation
of the salience maps (e.g., image gradients of expected cluster probabilities).}

\item{nMonte_variational}{An integer specifying how many Monte Carlo iterations to use in the
calculation of the expected likelihood in each training step.}

\item{TfRecords_BufferScaler}{The buffer size used in \code{tfrecords} mode is \code{batchSize*TfRecords_BufferScaler}. Lower \code{TfRecords_BufferScaler} towards 1 if out-of-memory problems.}

\item{dataType}{String specifying whether to assume \code{"image"} or \code{"video"} data types.}
}
\value{
Returns a list consiting of \itemize{
\item \code{clusterTaus_mean} default
\item \code{clusterProbs_mean}. Estimated mean image effect cluster probabilities.
\item \code{clusterTaus_sigma}. Estimated cluster standard deviations.
\item \code{clusterProbs_lowerConf}. Estimated lower confidence for effect cluster probabilities.
\item \code{impliedATE}. Implied ATE.
\item \code{individualTau_est}. Estimated individual-level image-based treatment effects.
\item \code{transportabilityMat}. Transportability matrix withestimated cluster information.
\item \code{plottedCoordinates}. List containing coordinates plotted in salience maps.
\item \code{whichNA_dropped}. A vector containing observations dropped due to missingness.
}
}
\description{
Implements the image heterogeneity decomposition analysis of Jerzak, Johansson, and Daoud (2023). Users
input in treatment and outcome data, along with a function specifying how to load in images
using keys referenced to each unit (since loading in all image data will usually not be possible due to memory limitations).
This function by default performs estimation, constructs salience maps, and can optionally perform
estimation for new areas outside the original study sites in a transportability analysis.
}
\section{References}{

\itemize{
\item Connor T. Jerzak, Fredrik Johansson, Adel Daoud. Image-based Treatment Effect Heterogeneity. Forthcoming in \emph{Proceedings of the Second Conference on Causal Learning and Reasoning (CLeaR), Proceedings of Machine Learning Research (PMLR)}, 2023.
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
