% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_BuildBackend.R
\name{BuildBackend}
\alias{BuildBackend}
\title{Perform causal estimation under image confounding}
\usage{
BuildBackend(conda_env, conda)
}
\arguments{
\item{conda_env}{(default = \code{"CausalImagesEnv"}) Name of the conda environment in which to place the backends.}

\item{conda}{(default = \code{auto}) The path to a conda executable. Using \code{"auto"} allows reticulate to attempt to automatically find an appropriate conda binary.}
}
\value{
Builds the computational environment for \code{causalimages}. This function requires an Internet connection.
You may find out a list of conda Python paths via: \code{system("which python")}
}
\description{
Perform causal estimation under image confounding
}
\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
