% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_WriteTfRecord.R
\name{GetElementFromTfRecordAtIndices}
\alias{GetElementFromTfRecordAtIndices}
\title{Reads indices from a \code{.tfrecord} file.}
\usage{
GetElementFromTfRecordAtIndices(indices, file)
}
\arguments{
\item{indices}{(integer vector) Observation indices to be retrieved from a \code{.tfrecord}}

\item{conda_env}{(Default = \code{NULL}) A \code{conda} environment where tensorflow v2 lives. Used only if a version of tensorflow is not already active.}

\item{conda_env_required}{(default = \code{F}) A Boolean stating whether use of the specified conda environment is required.}

\item{file}{(character string) A character string stating the path to a \code{.tfrecord}}
}
\value{
Returns content from a \code{.tfrecord} associated with \code{indices}
}
\description{
Reads indices from a \code{.tfrecord} file saved via a call to \code{causalimages::WriteTfRecord}. Assumes a tensorflow environment has been activated in R.
}
\examples{
# Example usage:
GetElementFromTfRecordAtIndices(
  indices = 1:10,
  file = "./NigeriaConfoundApp.tfrecord")

}
