% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_ImageModelBackbones.R
\name{GetImageRepresentations}
\alias{GetImageRepresentations}
\title{Generates image and video representations useful in earth observation tasks for casual inference, following the approach in Rolf, Esther, et al.  (2021).}
\usage{
GetImageRepresentations(imageKeysOfUnits, file, nWidth_ImageRep,
 conda_env, conda_env_required, kernelSize,
 temporalKernelSize, nWidth_ImageRep,
 strides, InitImageProcess, batchSize,
 TfRecords_BufferScaler, seed, dataType)
}
\arguments{
\item{imageKeysOfUnits}{A vector of length \code{length(imageKeysOfUnits)} specifying the unique image ID associated with each unit. Samples of \code{imageKeysOfUnits} are fed into the package to call images into memory.}

\item{file}{(default = \code{NULL}) Path to a tfrecord file generated by \code{WriteTfRecord}.}

\item{conda_env}{(default = \code{"CausalImagesEnv"}) A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}}

\item{conda_env_required}{(default = \code{T}) A Boolean stating whether use of the specified conda environment is required.}

\item{InitImageProcess}{(default = \code{NULL})}

\item{nWidth_ImageRep}{(default = \code{128L}) Number of embedding features output.}

\item{batchSize}{(default = \code{16L}) Integer specifying batch size in obtaining representations.}

\item{strides}{(default = \code{2L}) Integer specifying the strides used in the convolutional layers.}

\item{temporalKernelSize}{(default = \code{2L}) Dimensions used in the temporal part of the convolution kernels if using image sequences.}

\item{kernelSize}{(default = \code{5L}) Dimensif(ions used in the convolution kernels.}

\item{TfRecords_BufferScaler}{(default = \code{10L}) The buffer size used in \code{tfrecords} mode is \code{batchSize*TfRecords_BufferScaler}. Lower \code{TfRecords_BufferScaler} towards 1 if out-of-memory problems.}

\item{dataType}{(default = \code{"image"}) String specifying whether to assume \code{"image"} or \code{"video"} data types.}

\item{seed}{(default = \code{NULL}) Integer specifying the seed for pseudo random number generation.}
}
\value{
A list containing two items:
\itemize{
\item \code{Representations} (matrix) A matrix containing image/video representations, with rows corresponding to observations.
\item \verb{ImageRepArm_OneObs,ImageRepArm_batch_R, ImageRepArm_batch} (functions) Image modeling functions.
\item \code{ImageModel_And_State_And_MPPolicy_List} List containing image model parameters fed into functions.
}
}
\description{
Generates image and video representations useful in earth observation tasks for casual inference, following the approach in Rolf, Esther, et al.  (2021).
}
\section{References}{

\itemize{
\item Rolf, Esther, et al. "A generalizable and accessible approach to machine learning with global satellite imagery." \emph{Nature Communications} 12.1 (2021): 4392.
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
