% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_ImageModelBackbones.R
\name{GetImageRepresentations}
\alias{GetImageRepresentations}
\title{Generates image and video representations useful in earth observation tasks for casual inference.}
\usage{
GetImageRepresentations(
  imageKeysOfUnits = NULL,
  file = NULL,
  conda_env = "CausalImagesEnv",
  conda_env_required = T,
  returnContents = T,
  getRepresentations = T,
  imageModelClass = "VisionTransformer",
  Sys.setenv_text = NULL,
  InitImageProcess = NULL,
  nWidth_ImageRep = 64L,
  nDepth_ImageRep = 1L,
  nDepth_TemporalRep = 1L,
  batchSize = 16L,
  optimizeImageRep = T,
  strides = 1L,
  temporalKernelSize = 2L,
  kernelSize = 3L,
  patchEmbedDim = 16L,
  TfRecords_BufferScaler = 10L,
  dropoutRate,
  dataType = "image",
  bn_momentum = 0.9,
  inputAvePoolingSize = 1L,
  seed = NULL
)
}
\arguments{
\item{imageKeysOfUnits}{A vector of length \code{length(imageKeysOfUnits)} specifying the unique image ID associated with each unit. Samples of \code{imageKeysOfUnits} are fed into the package to call images into memory.}

\item{file}{Path to a tfrecord file generated by \code{causalimages::WriteTfRecord}.}

\item{conda_env}{A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}. Default = \code{"CausalImagesEnv"}}

\item{conda_env_required}{A Boolean stating whether use of the specified conda environment is required.}

\item{InitImageProcess}{(default = \code{NULL}) Initial image processing function. Usually left \code{NULL}.}

\item{nWidth_ImageRep}{Number of embedding features output.}

\item{batchSize}{Integer specifying batch size in obtaining representations.}

\item{strides}{Integer specifying the strides used in the convolutional layers.}

\item{temporalKernelSize}{Dimensions used in the temporal part of the convolution kernels if using image sequences.}

\item{kernelSize}{Dimensif(ions used in the convolution kernels.}

\item{TfRecords_BufferScaler}{The buffer size used in \code{tfrecords} mode is \code{batchSize*TfRecords_BufferScaler}. Lower \code{TfRecords_BufferScaler} towards 1 if out-of-memory problems.}

\item{dataType}{String specifying whether to assume \code{"image"} or \code{"video"} data types. Default is \code{"image"}.}

\item{seed}{Integer specifying the seed for pseudo random number generation.}
}
\value{
A list containing two items:
\itemize{
\item \code{Representations} (matrix) A matrix containing image/video representations, with rows corresponding to observations.
\item \verb{ImageRepArm_OneObs,ImageRepArm_batch_R, ImageRepArm_batch} (functions) Image modeling functions.
\item \code{ImageModel_And_State_And_MPPolicy_List} List containing image model parameters fed into functions.
}
}
\description{
Generates image and video representations useful in earth observation tasks for casual inference, following the approach in Rolf, Esther, et al.  (2021).
}
\section{References}{

\itemize{
\item Rolf, Esther, et al. "A generalizable and accessible approach to machine learning with global satellite imagery." \emph{Nature Communications} 12.1 (2021): 4392.
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
