% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_image2.R
\name{image2}
\alias{image2}
\title{Visualizing matrices as heatmaps with correct north-south-east-west orientation}
\usage{
image2(
  x,
  xaxt = NULL,
  yaxt = NULL,
  xlab = "",
  ylab = "",
  main = NULL,
  cex.main = NULL,
  col.lab = "black",
  col.main = "black",
  cex.lab = 1.5,
  box = F
)
}
\arguments{
\item{x}{The numeric matrix to be visualized.}

\item{xaxt}{The x-axis tick labels.}

\item{yaxt}{The y-axis tick labels.}

\item{xlab}{The x-axis labels.}

\item{ylab}{The y-axis labels.}

\item{main}{The main figure label.}

\item{cex.main}{The main figure label sizing factor.}

\item{col.lab}{Axis label color.}

\item{col.main}{Main label color.}

\item{cex.lab}{Cex for the labels.}

\item{box}{Draw a box around the image?}
}
\value{
Returns a heatmap representation of the matrix, \code{x}, with correct north/south/east/west orientation.
}
\description{
A function for generating a heatmap representation of a matrix with correct spatial orientation.
}
\examples{
#set seed
set.seed(1)

#Geneate data
x <- matrix(rnorm(50*50), ncol = 50)
diag(x) <- 3

# create plot
image2(x, main = "Example Text", cex.main = 2)

}
