% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_helperFxns.R
\name{image2}
\alias{image2}
\title{Visualizing matrices as heatmaps with correct north-south-east-west orientation}
\usage{
image2( x )
}
\arguments{
\item{x}{(required) The numeric matrix to be visualized.}

\item{xaxt}{(default = \code{""}) The x-axis tick labels.}

\item{yaxt}{(default = \code{""}) The y-axis tick labels.}

\item{xlab}{(default = \code{""}) The x-axis labels.}

\item{ylab}{(default = \code{""}) The y-axis labels.}

\item{main}{(default = \code{""}) The main figure label.}

\item{cex.main}{(default = \code{1.}) The main figure label sizing factor.}

\item{box}{(default = \code{F}) Should a box be plotted around the image?}
}
\value{
Returns a heatmap representation of the matrix, \code{x}, with correct north/south/east/west orientation.
}
\description{
A function for generating a heatmap representation of a matrix with correct spatial orientation.
}
\examples{
#set seed
set.seed(1)

#Geneate data
x <- matrix(rnorm(50*50), ncol = 50)
diag(x) <- 3

# create plot
image2(x, main = "Example Text", cex.main = 2)

}
