% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_helperFxns.R
\name{LongLat2CRS}
\alias{LongLat2CRS}
\title{Get the spatial point of long/lat coordinates}
\usage{
LongLat2CRS(long, lat, CRS_ref)
}
\arguments{
\item{long}{Vector of numeric longitudes.}

\item{lat}{Vector of numeric latitudes.}

\item{CRS_ref}{A CRS into which the long-lat point should be projected.}
}
\value{
Returns the long/lat location as a spatial point in the new CRS defined by \code{CRS_ref}
}
\description{
A function converts long/lat coordinates into a spatial points object defined by a coordinate reference system (CRS).
}
\examples{
spatialPt <- LongLat2CRS(
                  long = 49.932,
                  lat = 35.432,
                  CRS_ref = sp::CRS("+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +ellps=WGS84"))
}
