% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_helperFxns.R
\name{print2}
\alias{print2}
\title{print2}
\usage{
print2(text, quiet = F)
}
\arguments{
\item{x}{Character string to be printed, with date and time.}
}
\value{
Prints with date and time.
}
\description{
A function prints a string with date and time.
}
\examples{
print2("Hello world")
}
