% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_TrainDo.R
\name{TrainDo}
\alias{TrainDo}
\title{Runs a training routine (internal function)}
\usage{
TrainDo()
}
\arguments{
\item{.}{No parameters.}
}
\value{
Internal function performing model training.
}
\description{
Runs trainers defined in TrainDefine(). Internal function.
}
