% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_TrainDo.R
\name{TrainDo}
\alias{TrainDo}
\title{Runs a trainer.}
\usage{
TrainDo()
}
\arguments{
\item{.}{No parameters.}
}
\value{
Performs training.
}
\description{
Runs trainers defined in TrainDefine().
}
\section{References}{

\itemize{
\item Connor T. Jerzak, Fredrik Johansson, Adel Daoud. Image-based Treatment Effect Heterogeneity. Forthcoming in \emph{Proceedings of the Second Conference on Causal Learning and Reasoning (CLeaR), Proceedings of Machine Learning Research (PMLR)}, 2023.
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/causalimages-software/

}
