% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_TfRecordFxns.R
\name{WriteTfRecord}
\alias{WriteTfRecord}
\title{Write an image corpus as a .tfrecord file}
\usage{
WriteTfRecord(file, imageKeysOfUnits, acquireImageFxn,
conda_env, conda_env_required, writeVideo)
}
\arguments{
\item{file}{A character string naming a file for writing.}

\item{uniqueImageKeys}{A vector specifying the unique image keys of the corpus. A key grabs an image/video array via acquireImageFxn(key)}

\item{acquireImageFxn}{A function whose input is an observation index and whose output is an image.}

\item{writeVideo}{(default = \code{FALSE}) Should we assume we're writing image sequence data of form batch by time by height by width by channels?}

\item{conda_env}{(default = \code{"CausalImagesEnv"}) A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}}

\item{conda_env_required}{(default = \code{T}) A Boolean stating whether use of the specified conda environment is required.}
}
\value{
Writes a unique key-referenced \code{.tfrecord} from an image/video corpus for use in image-based causal inference training.
}
\description{
Writes an image corpus to a \code{.tfrecord} file for rapid reading of images into memory for fast ML training.
}
\examples{
# Example usage (not run):
#WriteTfRecord(
#  file = "./NigeriaConfoundApp.tfrecord",
#  uniqueImageKeys = 1:n,
#  acquireImageFxn = acquireImageFxn)

}
