% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_WriteTfRecord.R
\name{WriteTfRecord}
\alias{WriteTfRecord}
\title{Write an image corpus as a .tfrecord file}
\usage{
WriteTfRecord(file,acquireImageRepFxn,conda_env)
}
\arguments{
\item{file}{A character string naming a file for writing.}

\item{imageKeys}{A vector specifying the image keys of the corpus. A key grabs an image via acquireImageRepFxn(key)}

\item{acquireImageRepFxn}{A function whose input is an observation index and whose output is an image.}

\item{conda_env}{A \code{conda} environment where tensorflow v2 lives.}
}
\value{
Writes an index-referenced \code{.tfrecord} from an image corpus for use in image-based causal inference training.
}
\description{
Writes an image corpus to a \code{.tfrecord} file for rapid reading of images into memory for fast ML training.
}
\examples{
# Example usage:
#WriteTfRecord(
#  file = "./NigeriaConfoundApp.tfrecord",
#  acquireImageRepFxn = acquireImageRepFxn,
#  conda_env = "tensorflow_m1")

}
