% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalImage_TfRecordFxns.R
\name{WriteTfRecord}
\alias{WriteTfRecord}
\title{Write an image corpus as a .tfrecord file}
\usage{
WriteTfRecord(file,imageKeysOfUnits,acquireImageFxn,conda_env,conda_env_required)
}
\arguments{
\item{file}{A character string naming a file for writing.}

\item{acquireImageFxn}{A function whose input is an observation index and whose output is an image.}

\item{writeVideo}{(default = \code{FALSE}) Should we assume we're writing image sequence data of form batch by time by height by width by channels?}

\item{conda_env}{A \code{conda} environment where tensorflow v2 lives. Used only if a version of tensorflow is not already active.}

\item{conda_env_required}{(default = \code{F}) A Boolean stating whether use of the specified conda environment is required.}

\item{imageKeysOfUnits}{A vector specifying the image keys of the corpus. A key grabs an image via acquireImageFxn(key)}
}
\value{
Writes an key- and index-referenced \code{.tfrecord} from an image corpus for use in image-based causal inference training.
}
\description{
Writes an image corpus to a \code{.tfrecord} file for rapid reading of images into memory for fast ML training.
}
\examples{
# Example usage (not run):
#WriteTfRecord(
#  file = "./NigeriaConfoundApp.tfrecord",
#  keys = 1:n,
#  acquireImageFxn = acquireImageFxn,
#  conda_env = "tensorflow_m1")

}
