% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_TfRecordFxns.R
\name{WriteTfRecord}
\alias{WriteTfRecord}
\title{Write an image corpus as a .tfrecord file}
\usage{
WriteTfRecord(
  file,
  uniqueImageKeys,
  acquireImageFxn,
  writeVideo = F,
  image_dtype = "float16",
  conda_env = "CausalImagesEnv",
  conda_env_required = T,
  Sys.setenv_text = NULL
)
}
\arguments{
\item{file}{A character string naming a file for writing.}

\item{uniqueImageKeys}{A vector specifying the unique image keys of the corpus.
A key grabs an image/video array via \code{acquireImageFxn(key)}.}

\item{acquireImageFxn}{A function whose input is an observation keys and whose output is an array with dimensions \verb{(length(keys), nSpatialDim1, nSpatialDim2, nChannels)} for images and  \verb{(length(keys), nTimeSteps, nSpatialDim1, nSpatialDim2, nChannels)} for image sequence data.}

\item{writeVideo}{(default = \code{FALSE}) Should we assume we're writing image sequence data of form batch by time by height by width by channels?}

\item{image_dtype}{String specifying image data type for storage. Default is \code{"float16"}.}

\item{conda_env}{(default = \code{"CausalImagesEnv"}) A \code{conda} environment where computational environment lives, usually created via \code{causalimages::BuildBackend()}}

\item{conda_env_required}{(default = \code{T}) A Boolean stating whether use of the specified conda environment is required.}

\item{Sys.setenv_text}{Optional string for setting environment variables before Python initialization.}
}
\value{
Writes a unique key-referenced \code{.tfrecord} from an image/video corpus for use in image-based causal inference training.
}
\description{
Writes an image corpus to a \code{.tfrecord} file for rapid reading of images into memory for fast ML training.
Specifically, this function serializes an image or video corpus into a \code{.tfrecord} file, enabling efficient data loading for machine learning tasks, particularly for image-based causal inference training.
It requires that users define an \code{acquireImageFxn} function that accepts keys and returns the corresponding image or video as an array of dimensions \verb{(length(keys), nSpatialDim1, nSpatialDim2, nChannels)} for images or \verb{(length(keys), nTimeSteps, nSpatialDim1, nSpatialDim2, nChannels)} for video sequences.
}
\examples{
# Example usage (not run):
#WriteTfRecord(
#  file = "./NigeriaConfoundApp.tfrecord",
#  uniqueImageKeys = 1:n,
#  acquireImageFxn = acquireImageFxn)

}
